Refactor cpp_dump macros to use static_cast for size calculations, enhancing type safety and consistency.

+++ cpp-dump.hpp
@@ -24,7 +24,7 @@
 
 #define _p_CPP_DUMP_STRINGIFY(x) #x
 #define _p_CPP_DUMP_CONTAINS_VARIADIC_TEMPLATE(...)                                                \
-  cpp_dump::_detail::contains_variadic_template<_p_CPP_DUMP_VA_SIZE(__VA_ARGS__)>(                 \
+  cpp_dump::_detail::contains_variadic_template<static_cast<std::size_t>(_p_CPP_DUMP_VA_SIZE(__VA_ARGS__))>( \
       {_p_CPP_DUMP_EXPAND_VA(_p_CPP_DUMP_STRINGIFY, __VA_ARGS__)}                                  \
   )
 
@@ -36,7 +36,7 @@
  */
 #define cpp_dump(...)                                                                              \
   cpp_dump::_detail::cpp_dump_macro<                                                               \
-      _p_CPP_DUMP_VA_SIZE(__VA_ARGS__),                                                            \
+      static_cast<std::size_t>(_p_CPP_DUMP_VA_SIZE(__VA_ARGS__)),                                  \
       _p_CPP_DUMP_CONTAINS_VARIADIC_TEMPLATE(__VA_ARGS__)>(                                        \
       {__FILE__, __LINE__, __func__},                                                              \
       {_p_CPP_DUMP_EXPAND_VA(_p_CPP_DUMP_STRINGIFY, __VA_ARGS__)},                                 \
